+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [0.1.0] - 31th of August 2020
 
+- Move examples into one file
+- Remove shadowing type parameters
+
+## [0.1.0] - 31th of August 2020
+
 - Create basic classes
 - Write documentation
 - Use [effective_dart](https://pub.dev/packages/effective_dart)
