+++ pubspec.yaml
@@ -1,6 +1,7 @@
 name: deep_collection
 description: Extends [List], [Set] and [Map] in order to add commonly used recursive methods such as sorting, reversing, finding intersections or differences.
-version: 0.1.0
+version: 0.1.1
+homepage: https://github.com/owczaro/deep_collection
 
 environment:
   sdk: '>=2.7.0 <3.0.0'
